backend/requirements.txt
flask
flask-restx
google-auth
python-dotenv
wrapt
